+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.29.2] - 2018-11-07
+### Fixed
+- remove autocomplete suggestions onBlur
+- show an initial list of values on focus
+
 ## [0.29.1] - 2018-11-01
 ### Fixed
 - Suggestions were not accessible by keydown event.